Report code coverage via CodeCov.

+++ deploy/clean.py
@@ -1,16 +1,31 @@
-﻿"""Deletes cache and build artifacts."""
-__license__ = 'MIT'
+﻿"""Deletes build and test artifacts."""
 
 from pathlib import Path
 from shutil import rmtree
 
 root = Path(__file__).resolve().parent.parent
 
-for folder in root.rglob('__pycache__'):
-    rmtree(folder, ignore_errors=True)
+folders = [
+    root/'deploy'/'docs',
+    root/'deploy'/'coverage',
+    root/'deploy'/'dist',
+    root/'dist',
+]
+folder_names = [
+    '__pycache__',
+    '.pytest_cache',
+]
+for folder_name in folder_names:
+    for folder in root.rglob(folder_name):
+        folders.append(folder)
+for folder in folders:
+    if folder.is_dir():
+        rmtree(folder, ignore_errors=True)
 
-for folder in root.rglob('.pytest_cache'):
-    rmtree(folder)
-
-for folder in ('docs', 'dist', 'coverage'):
-    rmtree(root/'deploy'/folder, ignore_errors=True)
+files = [
+    root/'.coverage',
+    root/'coverage.xml',
+]
+for file in files:
+    if file.is_file():
+        file.unlink()
